requirements.txt
apache-airflow==2.9.0
opencv-python-headless==4.9.0.80
moviepy==1.0.3
pytesseract==0.3.10
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
imageio
